requirements.txt
aws-cdk-lib==2.3.0
constructs>=10.0.0,<11.0.0
aws_cdk.aws_lambda_python_alpha
boto3
aws_lambda_powertools
requests
bs4
pytz